refactor(matrix): standardize matrix dimensions to u32 (#295)
* refactor(matrix): standardize matrix dimensions to u32

Change the internal storage of matrix dimensions (rows and cols) from
usize to u32 in both Matrix and SMatrix definitions. This requires
explicit casting in many call sites across the codebase.

* fix(fdm): prevent integer overflow when calculating image size

Adds checks in `setTarget` and `setSource` to ensure image dimensions do not
exceed the maximum size representable by u32, returning `error.Overflow`
if necessary.

Also updates initialization syntax to use explicit type annotations
(e.g., `var x: Type = .init(...)`) across the module and tests.

Ref: #32

* refactor(core): standardize matrix dimensions and indices to u32

Change all matrix dimensions, indices, ranks, and component counts from
platform-dependent usize to fixed-width u32 across the codebase.

* refactor(matrix): standardize dimension types and casting

Ensure matrix allocation size calculations use usize to prevent overflow.
Standardize Python index parsing to isize and update dimension validation
in Python bindings for consistency.

Ref: #32

* fix(bindings/python): increase error buffer size and improve type handling

Increase error message buffer size from 256 to 512 bytes to prevent
potential overflows during range validation formatting.

Also ensures `isize` is correctly handled in range validation conversion
and fixes a typo in the matrix column validation message ("col" -> "cols").

Ref: #32

+++ bindings/python/src/pca.zig
@@ -156,7 +156,7 @@ fn pca_fit(self_obj: ?*c.PyObject, args: ?*c.PyObject, kwds: ?*c.PyObject) callc
     }
 
     // Convert num_components to optional usize
-    const components: ?usize = if (num_components < 0) null else @intCast(num_components);
+    const components: ?u32 = if (num_components < 0) null else @intCast(num_components);
 
     // Fit the PCA model
     self.pca_ptr.?.fit(matrix_struct.matrix_ptr.?.*, components) catch |err| {
